Added missing in-bounds assumptions about the vault
We need to ensure that all vault fields are in-bounds.

The vault (from Solana account) is just uninitialized memory for the
prover. Then, the verification state FvVault is created from the
Solana vault but in-bounds information is not propagated, even if
FvVault::from converts a u64 to NativeInt. Unfortunately, this not
enough for the prover to know that FvVault fields are actually within
u64 bounds.